+++ test/GPU/primitive_dry.jl
@@ -1,6 +1,7 @@
 @testset "GPU PrimitiveDryModel" begin
-    spectral_grid = SpectralGrid(trunc=32, nlayers=8, architecture=SpeedyWeather.GPU())
-    model = CUDA.@allowscalar PrimitiveDryModel(spectral_grid=spectral_grid)
+    arch = SpeedyWeather.GPU()
+    spectral_grid = SpectralGrid(trunc=32, nlayers=8, architecture=arch)
+    model = PrimitiveDryModel(spectral_grid=spectral_grid, physics=false)
     simulation = CUDA.@allowscalar initialize!(model)
     run!(simulation, steps=4)